postgres_to_es/requirements/base.txt
psycopg2-binary==2.9.1
pydantic==1.8.2
elasticsearch==7.15.0
django-environ==0.4.5